Fix tooltip and extract shared code
