+++ app/javascript/app/pages/country-context/socioeconomic-indicators/population/population-styles.scss
@@ -1,39 +1,2 @@
 @import '~styles/settings.scss';
 @import '~styles/layout.scss';
-
-$dropdown-min-width: 265px;
-
-.page {
-  @include row();
-
-  margin-top: $gutter-padding;
-}
-
-.container {
-  @include columns($columns: 12);
-
-  @media #{$tablet-landscape} {
-    padding-left: $gutter-padding / 2;
-
-    @include columns($columns: (6,6));
-  }
-}
-
-.toolbox {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 20px;
-}
-
-.dropdown {
-  max-width: $dropdown-min-width;
-}
-
-.buttonWrapper {
-  align-self: flex-end;
-  max-width: 100px;
-}
-
-.legend {
-  margin-top: 15px;
-}
