+++ app/javascript/app/pages/country-context/socioeconomic-indicators/shared/bar-chart-tooltip/bar-chart-tooltip.js
