+++ app/javascript/app/pages/country-context/socioeconomic-indicators/shared/socioeconomic-selectors.js
